requirements/local.txt
-r base.txt

boto3==1.35.24
botocore==1.35.24
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
django-rest-authtoken==2.1.4
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.3
